LayoutTests/http/tests/referrer-policy-script/unsafe-url/cross-origin-http.https-expected.txt

Tests the behavior of unsafe-url referrer policy when cross origin from HTTPS to HTTP.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is "https://127.0.0.1:8443/referrer-policy-script/unsafe-url/cross-origin-http.https.html"
PASS successfullyParsed is true

TEST COMPLETE
